Add a flag for Darwin linking behaviour.
Darwin does want to be able to resolve all symbols during linking, but there's no need to bake everything into libomu.so, as we do in for windows. Instead, we can supply -lhqp during both libomu.so linking and executable linking. We overload the hqp_libs_resolved flag with the new content 'partial' to flag this behaviour.